extrapolate RoomAnalytics into its own service
